+++ CHANGELOG.md
@@ -5,6 +5,7 @@ To see version information, `xsvutils --version`.
 To see compatibility information, `xsvutils help compatibility`.
 
 ## version 0.3.4 (Unreleased)
+- add `ratio` subcommand
 
 ## version 0.3.3 (2018/05/27)
 - add `rmnoname` subcommand